Add TypeScript transformer and move a bunch of tests over
There are still two failing tests that are marked as skipped, but most of them
just worked. I also had to do the same hack as in flow where I muck with one of
the tokens to denote a type parameter start.
