+++ test/util.ts
@@ -5,7 +5,7 @@ import {transform, Transform} from "../src";
 export function assertResult(
   code: string,
   expectedResult: string,
-  transforms: Array<Transform> = ["jsx", "imports", "react-display-name", "flow"],
+  transforms: Array<Transform> = ["jsx", "imports", "react-display-name"],
 ): void {
   assert.equal(transform(code, {transforms}), expectedResult);
 }
