+++ website/src/getTokens.js
@@ -1,8 +1,8 @@
 import {getFormattedTokens} from 'sucrase';
 
-export default function getTokens(code) {
+export default function getTokens(code, transforms) {
   try {
-    return getFormattedTokens(code)
+    return getFormattedTokens(code, {transforms})
   } catch (e) {
     console.error(e);
     return e.message;
